requirements.txt
asgiref==3.9.1
astroid==3.3.11
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
autopep8==2.3.2
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
Deprecated==1.2.18
dill==0.4.0
Django==5.2.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flake8==7.3.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isort==6.0.1
lazy-object-proxy==1.12.0
mccabe==0.7.0
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.10.1
pylint==3.3.8
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
stripe==12.5.1
tomli==2.2.1
tomlkit==0.13.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
urllib3==2.5.0
wget==3.2
wrapt==1.17.3
zope.interface==7.2
